feat: enhance expert application process and update admin interface
- Removed portfolio URL field from the expert application admin interface.
- Implemented user role update to 'Expert' upon application approval in the review process.
- Updated the admin dashboard to display pending expert applications count.
- Refactored expert application data handling in the frontend to improve user experience and error handling.
- Simplified the analytics page by removing unused chart configurations.

+++ frontend-web/src/utils/network-optimization.ts
@@ -180,7 +180,7 @@ export async function compressData(data: string): Promise<Uint8Array> {
     while (true) {
       const { done, value } = await reader.read()
       if (done) break
-      chunks.push(value)
+      chunks.push(value as Uint8Array)
     }
 
     // Combine chunks
